moved all other files into respective folders
