+++ lib/src/Dialog.js
@@ -1,4 +1,4 @@
-import DialogNode from "./DialogNode";
+import DialogNode from "./nodes/DialogNode";
 
 export default class Dialog {
   /**
