+++ lib/src/contents/DialogReferenceContent.js
