+++ lib/src/contents/DialogTextContent.js
