+++ lib/src/converters/convertDataToDialog.js
@@ -1,7 +1,7 @@
 import Dialog from "../Dialog";
-import DialogNode from "../DialogNode";
-import DialogTextContent from "../DialogTextContent";
-import DialogReferenceContent from '../DialogReferenceContent';
+import DialogNode from "../nodes/DialogNode";
+import DialogTextContent from "../contents/DialogTextContent";
+import DialogReferenceContent from '../contents/DialogReferenceContent';
 
 /**
  * Will convert data from server to a dialog.
