+++ lib/src/index.js
@@ -1,8 +1,8 @@
-import convertDataToDialog from './convertDataToDialog';
+import convertDataToDialog from './converters/convertDataToDialog';
 import Dialog from './Dialog';
-import DialogNode from './DialogNode';
-import DialogTextContent from './DialogTextContent';
-import DialogReferenceContent from "./DialogReferenceContent"
+import DialogNode from './nodes/DialogNode';
+import DialogTextContent from './contents/DialogTextContent';
+import DialogReferenceContent from "./contents/DialogReferenceContent"
 
 export {
     convertDataToDialog,
