+++ lib/src/nodes/DialogNode.js
@@ -1,4 +1,5 @@
-import DialogTextContent from "./DialogTextContent";
+import DialogTextContent from "../contents/DialogTextContent";
+
 export default class DialogNode {
   /**
    * @param {String} id
