+++ lib/tests/Dialog.spec.js
@@ -1,6 +1,4 @@
-import Dialog from "./Dialog";
-import DialogNode from "./DialogNode";
-import DialogTextContent from "./DialogTextContent";
+import { Dialog, DialogNode, DialogTextContent } from "../src";
 
 const testContent = new DialogTextContent(
   "9BEB03C5-867C-4355-AEBD-D3A64C82FFAA",
