+++ lib/tests/DialogNode.spec.js
@@ -1,5 +1,4 @@
-import DialogNode from "./DialogNode";
-import DialogTextContent from "./DialogTextContent";
+import { DialogNode, DialogTextContent } from "../src";
 
 const testContent = new DialogTextContent(
   "9BEB03C5-867C-4355-AEBD-D3A64C82FFAA",
