+++ lib/tests/DialogReferenceContent.spec.js
@@ -1,4 +1,5 @@
-import DialogReferenceContent from "./DialogReferenceContent";
+import { DialogReferenceContent } from "../src";
+
 
 describe("DialogReferenceContent.ctor", () => {
     it("correct data, should create", () => {
