+++ lib/tests/DialogTextContent.spec.js
@@ -1,4 +1,4 @@
-import DialogTextContent from "./DialogTextContent";
+import { DialogTextContent } from "../src";
 
 describe("DialogTextContent.ctor", () => {
   it("correct data, should create", () => {
